0.2.0
Thats right, 0.2.0!

Click and Dig

Summary:
1: Can now traverse the stars with a GUI
2: Can now Mine the crusts of planets
3: Each ship can now contain items (eg. Iron, Hydrogen, etc)
4: Systems now have better names. You can review in formulas.py

1: GUI!!!
Honestly something ive wanted since i started this.
So far the only things that can be done are travelling to other solar systems and planets within the solar system, along with getting information and mining planets. Also you can switch to the Console

You can switch to the console by clicking the console button on the top right
You can travel to stars or planets using the Travel button on the top left
If you are in the console, you can return to the GUI by typing exit or /exit
If you are in the console and it gets confused when you press a button, just press enter, it has not crashed in my testing.

2: Mining!!
You can now visit a planet and mine from the crust of the planet (or atmosphere if it is a gas giant). Information on the math part of it is below. The TLDR is the most abundant element has a chance of entering your inventory when you mine for it.

3: Inventory!...
wait wasnt this already here? Nope, there was a number for it but there were no specific inventory slots. Now there are! The slots are all available elements in the game, although there isnt much to do yet... yet...

4: System Names...:
Okay yes we did have this, I know. Now there are words! It first takes a word from the nouns list and adds a word from the adjectives list, followed by a letter from the English Alphabet. There are around 3,000 system names ignoring letters. (With letters its 78,000... please dont make 78,000 systems)

Can now mine the crusts/atmospheres of planets. This takes the value of the most abundant element in the crust and does a random calculation based on that value if you recieve any of the element.

For example, lets say the crust of a rocky world has 40% Iron. You must roll at or below 40 to collect iron (this rolls out of 100). It will then take that roll and if it is below or equal to 40%, it will divide by 10 and that is the amount of the resource you gain.

Commands added:
/planet scan   Will provide the most abundant resource in Crust or Atmosphere
/planet mine   Will mine the most abundant element and add it to the inventory. Mining will fail once you reach the Hold limit
/exit   Will exit the console and return to the GUI

Inventories now exist. Ships can have any element that exists here, but starts at 0.