0.2.2
A Stellar Update!

Thats right, this is a super cool update for Stars

--Remember to Create a new galaxy, type the following command in the console:
/make galaxy 100 True
(Where 100 is the minimum number of systems, and True is whether you want moons)

This also adjusts solar system creation to account for the startype
(eg. Red Dwarfs will have a lot less Gas giants than a G-Type Star)

Also included are the possibility of multiple stars in a system! (Because fun fact, most star systems have multiple stars)
The highest it can go is 7 Stars in a system (as that is the highest seen irl, but the chances of it getting that high is pretty rare, like it would have to be only red dwarfs with a Blue Giant

Stellar Properties include the following:
Class: The Class, eg. O, B, A, F, G, K, M
Values: This is the number of points allowed for planets
Abundance: This determines the likelyhood of the star being chosen
Radius: The Radius
Mass: The Mass
Color: The Color
Temperature: The Temperature

Star Chances (not realistic, if it was the hope of seeing a blue giant would be nothing, and all youd actually see are a bunch of red dwarfs)

O-Class Stars: 1%
B-Class Stars: 3%
A-Class Stars: 6%
F-Class Stars: 10%
G-Class Stars: 20% (This is our Sun, and the starting Star)
K-Class Stars: 20%
M-Class Stars: 40%

Starting System Update
-The Starting System Star is now always a G-Class Star
-The Planets are now ordered as 4 Rocky, 2 Gas Giants, and 2 Ice Giants
-There is no room for a second star in the Starting System
-All planets except for the third one are forced to be uninhabitable

Adjustments to moon creation:
This will be expanded on in the future, but the possibility to now only generate a single moon at a time has been implemented. This is only used with the /make command at this point, but it is my goal to adjust the whole moon generation to happen outside of the moonfactory function and have it at the planet level.

Fixes:
1: Overflow systems now are given planets of their own
2: the systems.json file that stores planet data is no longer sorted alphabeticaly, but now sorted by properties
3: `/make moon` now returns a moon, which it seems it wasnt doing before. It will also now print a single moon, instead of multiple. Also added is the ability to save your moon as a file labelled myMoon.json

Development is being worked on to get functioning Space Stations, I havent forgotten!!