0.2.3
Save Games!

Yes you heard me right! Files are now saved. The files saved include the player data and galaxy data. It is also now possible to reload to a save file should you need it

Also possible is the ability to generate a new galaxy within the GUI!

All these are available under the Settings Button

Also on startup, it now loads the player.json file (Previously it only loaded the systems.json, since that was the only file to load)

Fixes:
Renamed the ship type Diplomatic to Diplomat. a "Massive Diplomatic" just doesnt sound right in my mind.