0.2.5
Another GUI Update!

Added the commands /save and /load

You will now be prompted to create a new galaxy should the
savefile fail to load

Removed use of tkinter.tkk in favor of just tkinter

Console Moved to the GUI. The Python console is now completely unusable (I wanted to keep the console as an option, but frankly it was slowing me down a bit, and being able to fully move every feature to a UI console as the cost of input() I feel is worth it.

Made the Grid easier to view due to it not being destroyed every time it is called

Element Grid now lines up more with the buttons that previously

A solution has been worked on to Adjust screen size as needed. Currently I have reduced the size of the working parts of the screen to something that should be usable to all