0.2.6
AI Ships!

As I get closer to Stations, Here is the addition of AI Ships!
Are the smart? Not at all, but they are a work in progress. AI Ships are not currently saved to the game when a game is saved

So far they can travel to star systems. Thats it. Literally all they do is fly around. But it uses a new file, called SystemLinks.json. This file has a full list of the quickest routes to each star system from each star system.

Using this file, the AI Ships can decide on a destination and will know where to travel and how to get there.
The creation of the file depends on the Map size.
Load Times:
Large Galaxy: 15 Seconds
Medium Galaxy: 3 Seconds
Small Galaxy: <1 Second

Also shown are tabs for the Screen!
There are currently 2:
-Console
-Ship View

Ship View will show all Ships in the current system, along with its ID, Ship Type, and Allegiance

Yes! There are now 4 Nations, but so far all they are only ship properties (youd imagine id make the nations first and add them to ships after, but nope)
These Nations are:
-Solar Empire
-Republic
-Pirates
-Coalition

Soon there will be descriptions of them all